+++ python/GafferOSLUI/OSLCodeUI.py
@@ -259,7 +259,7 @@ class _CodePlugValueWidget( GafferUI.PlugValueWidget ) :
 
 	def __init__( self, plug, **kw ) :
 
-		self.__codeWidget = GafferUI.CodeWidget()
+		self.__codeWidget = GafferUI.CodeWidget( lineNumbersVisible=True )
 
 		GafferUI.PlugValueWidget.__init__( self, self.__codeWidget, plug, **kw )