+++ Changes.md
@@ -9,7 +9,12 @@ Improvements
 - ColorChooser :
   - Changed the color field widget to a color wheel when hue is one of the varying components. [^1]
   - Changed the indicator for the color field and color sliders to an unfilled circle so the chosen color is visible in the center.
-- Python Editor : Added line numbers (#6091).
+- PythonEditor, PythonCommand, Expression, UIEditor, OSLCode : Added line numbers to code editors (#6091).
+
+Fixes
+-----
+
+- Expression, OSLCode : Fixed line numbers reported in OSL parse errors.
 
 API
 ---